requirements.txt
Django==5.0.1
djangorestframework==3.14.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
psycopg2-binary==2.9.6
python-decouple==3.8
django-storages==1.14.2
boto3==1.34.19
Pillow==10.2.0
drf-yasg==1.21.7
django-filter==23.5